Moved GetCourseType method to Dfc.CourseDirectory.Core project.
