+++ src/Dfc.CourseDirectory.Core/Services/ICourseTypeService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Dfc.CourseDirectory.Core.Services
+{
+    public interface ICourseTypeService
+    {
+        public Task<Models.CourseType?> GetCourseType(string learnAimRef);
+    }
+}
