--- tests/Dfc.CourseDirectory.Web.Tests/MockHttpSession.cs
@@ -1,47 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-
-namespace Dfc.CourseDirectory.Web.Tests
-{
-    public class MockHttpSession : ISession
-    {
-        readonly Dictionary<string, object> _sessionStorage = new Dictionary<string, object>();
-        string ISession.Id => throw new NotImplementedException();
-        bool ISession.IsAvailable => throw new NotImplementedException();
-        IEnumerable<string> ISession.Keys => _sessionStorage.Keys;
-        void ISession.Clear()
-        {
-            _sessionStorage.Clear();
-        }
-        Task ISession.CommitAsync(CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-        Task ISession.LoadAsync(CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-        void ISession.Remove(string key)
-        {
-            _sessionStorage.Remove(key);
-        }
-        void ISession.Set(string key, byte[] value)
-        {
-            _sessionStorage[key] = Encoding.UTF8.GetString(value);
-        }
-        bool ISession.TryGetValue(string key, out byte[] value)
-        {
-            if (_sessionStorage[key] != null)
-            {
-                value = Encoding.ASCII.GetBytes(_sessionStorage[key].ToString());
-                return true;
-            }
-            value = null;
-            return false;
-        }
-    }
-}
